Add welcome message
Currently there is no display of messages when user starts the
program.

Users may feel lost and do not know what to do next.

Add a welcome message function in MainWindow and Paimon class.

This ensures there is always a welcome message display from MainWindow
class and gives more flexibility if the welcome message has to be
changed in the future.

(attempt to write full commit message)