MainWindow class: Bind vbox height to scrollPane
Currently after users type a msg, the chat does not
automatically scroll to the latest conversation.

This might be inconvenient for users as they need to keep
scrolling their conversations down.

A Listener has been added to the height property of VBox. Whenever
the height increases, the scrollPane will reset its height back
to 1.0

The height property of VBox has been bind to the scrollPane
viewport height property.

(attempt to write a full commit msg)