Remove trailing and leading whitespace in user commands
Currently commands like " list" or "list " are recognised as
invalid commands.

This is not ideal as users may be confused after following
the intructions to type the commands.

Trim in user input in Parse.parse function.

Adding it as part of the command validation code block ensures
all syntax checkings are at one place. It is also neater if
Parser class handles all the syntax checks before passing the strings
to the downstream classes.