Check empty command parameters
Currently if any any command parameter is empty, the program
quits with runtime error.

This should be handled explicitly and the program should remind
user the correct way of keying in the command parameters.

Throws PaimonInvalidInputException if the command parameter is empty.

This is done in each parsing function. Although this is a bit
repetitive, it is necessary as we do not long the length of
command and required parameters without first parsing them.

(attempt to write full commit msg)